Lv5 work

+++ app/src/main/java/com/example/myloginpage/User.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "user")
 data class User(
     @PrimaryKey(autoGenerate = true) //키값을 부여할것이 필요
-    val userIndex : Long = 0L,
+    val userIndex : Int = 0,
     @ColumnInfo //데이터에서 행으로 쓸꺼면 필요, pw는 필요없을지도?
     var id: String,
     @ColumnInfo
